Move to configuration service, update actions/services name  - #3757
