--- ui/app/src/services/locale.ts
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2007-2021 Crafter Software Corporation. All Rights Reserved.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as published by
- * the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
-import { fetchConfigurationDOM } from './configuration';
-import { deserialize } from '../utils/xml';
-
-export function fetchLocaleSettings(site: string): Observable<any> {
-  return fetchConfigurationDOM(site, '/site-config.xml', 'studio').pipe(
-    map((xml) => {
-      let settings = {};
-      if (xml) {
-        const localeXML = xml.querySelector('locale');
-        if (localeXML) {
-          settings = deserialize(localeXML).locale;
-        }
-      }
-      return settings;
-    })
-  );
-}
