+++ ui/app/src/state/actions/system.ts
@@ -105,6 +105,6 @@ export const fetchGlobalMenuComplete = /*#__PURE__*/ createAction<GlobalState['u
 
 export const fetchGlobalMenuFailed = /*#__PURE__*/ createAction('FETCH_GLOBAL_MENU_FAILED');
 
-export const fetchLocaleSettings = /*#__PURE__*/ createAction('FETCH_LOCALE_SETTINGS');
-export const fetchLocaleSettingsComplete = /*#__PURE__*/ createAction('FETCH_LOCALE_SETTINGS_COMPLETE');
-export const fetchLocaleSettingsFailed = /*#__PURE__*/ createAction('FETCH_LOCALE_SETTINGS_FAILED');
+export const fetchSiteLocale = /*#__PURE__*/ createAction('FETCH_SITE_LOCALE');
+export const fetchSiteLocaleComplete = /*#__PURE__*/ createAction('FETCH_SITE_LOCALE_COMPLETE');
+export const fetchSiteLocaleFailed = /*#__PURE__*/ createAction('FETCH_SITE_LOCALE_FAILED');
